backend/requirements.txt
fastapi
uvicorn>=0.23.2
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic-settings>=2.0.3
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0.post1
bcrypt>=4.0.1
boto3>=1.28.64
stripe>=7.3.0
sendgrid>=6.10.0
redis>=5.0.1
celery>=5.3.4
pytest>=7.4.3
httpx>=0.25.1
